# apps/web-h5/docs/apifox/feedback.openapi.yaml
openapi: 3.0.3
info:
  title: GO-TECH 需求反馈接口
  version: 1.1.0
  description: |
    GO-TECH 需求反馈模块接口，可直接导入 Apifox。

    前台写操作使用 `GO_TECH_AUTH_TOKEN` Cookie 登录态；后台管理接口使用 Bearer JWT。
    新增建议和评论还需要 Cloudflare Turnstile Token，或备用验证码组合 Token。

    AI 助手的公开入口仍由 `web-h5` 提供：`web-h5` 完成管理员鉴权，并将请求代理到内部
    Python `feedback-ai` 服务。浏览器不得直接访问 Python 服务，也不需要携带内部服务 Token。
servers:
  - url: http://localhost:3001
    description: 本地 web-h5 开发环境
...
paths:
...
  /api/feedback/features:
    get:
      tags: [反馈-公共]
      summary: 搜索反馈建议
      description: 按标题或内容模糊搜索未删除建议，按创建时间倒序返回，最多 50 条；搜索词为空时返回空数组。
      operationId: searchFeedbackFeatures
      parameters:
        - name: q
          in: query
          required: false
          description: 标题或建议内容搜索关键词。
          schema:
            type: string
            example: 发票
        - name: system
          in: query
          required: false
          description: 按所属系统筛选搜索结果；省略时搜索全部系统。
          schema:
            type: string
            enum: [pms, hr]
            example: pms
      responses:
        '200':
          description: 建议搜索成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureListResponse'
    post:
      tags: [反馈-用户]
      summary: 新增反馈建议
      description: 新增建议并自动为提交人投 1 票；提交前执行机器人验证和内容安全审核。
      operationId: createFeedbackFeature
      security:
        - UserCookieAuth: []
      requestBody:
        required: true
        description: 新增建议所需的分类、标题、内容及机器人验证凭证。
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeatureRequest'
      responses:
        '201':
          description: 建议创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UserUnauthorized'
        '422':
          $ref: '#/components/responses/ContentRejected'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    UserCookieAuth:
      type: apiKey
      in: cookie
      name: GO_TECH_AUTH_TOKEN
      description: web-h5 用户登录后由浏览器自动携带的 JWT Cookie。
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 后台管理员登录后取得的 HS512 JWT，放入 Authorization 请求头。
...
  responses:
    BadRequest:
      description: 请求参数、请求体或机器人验证无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UserUnauthorized:
      description: 用户未登录或登录 Cookie 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ContentRejected:
      description: 标题、建议内容或评论未通过内容安全审核
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: 机器人验证、备用验证码或内容审核服务暂时不可用
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    FeatureStatus:
      type: string
      description: 建议处理状态；pending 为待评估，developing 为开发中，shipped 为已完成。
      enum: [pending, developing, shipped]
    ErrorResponse:
      type: object
      description: 接口调用失败时的统一响应。
      required: [success, message]
      properties:
        success:
          type: boolean
          description: 请求是否成功，错误响应固定为 false。
          example: false
        code:
          type: string
          nullable: true
          description: 可选的机器可读错误码，主要用于机器人验证和内容审核错误。
          enum: [BOT_VERIFICATION_FAILED, BOT_VERIFICATION_UNAVAILABLE, CONTENT_MODERATION_FAILED, CONTENT_REJECTED]
        message:
          type: string
          description: 面向用户或调用方的中文错误说明。
          example: 请求参数无效
...
    FeedbackComment:
      type: object
      description: 前台展示的普通评论或官方回复。
      required: [id, author, content, createdAt, isOfficial]
      properties:
        id:
          type: string
          description: 评论 ID。
          example: '2048'
        author:
          type: string
          description: 评论作者名称；官方回复固定为 GO-TECH Manager。
          example: GO-TECH Manager
        content:
          type: string
          description: 评论或回复正文。
          example: 该建议已进入评估阶段。
        createdAt:
          type: string
          format: date-time
          description: 评论创建时间，ISO 8601 格式。
        isOfficial:
          type: boolean
          description: 是否为官方回复。
          example: true
        parentId:
          type: string
          nullable: true
          description: 被回复的普通评论 ID；一级评论或建议级官方回复为 null。
          example: '1025'
    FeedbackFeature:
      type: object
      description: 前台反馈建议详情。
      required: [id, title, description, author, category, likes, likedBy, comments, status, createdAt]
      properties:
        id:
          type: string
          description: 建议 ID。
          example: '1024'
        title:
          type: string
          description: 建议标题。
          example: 增加批量导出发票功能
        description:
          type: string
          description: 建议详细内容。
          example: 希望订单页面可以按月份批量导出发票。
        author:
          type: string
          description: 建议提交人名称。
          example: 张先生
        category:
          type: string
          description: 一级分类名称。
          example: 系统功能
        subCategory:
          type: string
          nullable: true
          description: 子分类名称，未选择时为空。
          example: 订单管理
        likes:
          type: integer
          description: 当前建议票数。
          example: 12
        likedBy:
          type: array
          description: 已点赞用户名称列表。
          items:
            type: string
            description: 已点赞用户名称。
        comments:
          type: array
          description: 未删除且可见的评论和官方回复。
          items:
            $ref: '#/components/schemas/FeedbackComment'
        status:
          $ref: '#/components/schemas/FeatureStatus'
        createdAt:
          type: string
          format: date-time
          description: 建议创建时间，ISO 8601 格式。
        shippedAt:
          type: string
          format: date-time
          nullable: true
          description: 建议完成时间；未完成时为空。
        version:
          type: string
          nullable: true
          description: 建议发布或完成时关联的产品版本号。
          example: v2.3.0
    FeatureListResponse:
      type: object
      description: 前台建议列表响应。
      required: [success, data]
      properties:
        success:
          type: boolean
          description: 请求是否成功。
          example: true
        data:
          type: array
          description: 匹配的建议数组。
          items:
            $ref: '#/components/schemas/FeedbackFeature'
    FeatureResponse:
      type: object
      description: 单条建议操作成功响应。
      required: [success, data]
      properties:
        success:
          type: boolean
          description: 请求是否成功。
          example: true
        data:
          $ref: '#/components/schemas/FeedbackFeature'
    CreateFeatureRequest:
      type: object
      description: 新增反馈建议请求体。
      required: [category, title, description, turnstileToken]
      properties:
        category:
          type: string
          minLength: 1
          description: 已存在的一级分类名称。
          example: 系统功能
        subCategory:
          type: string
          nullable: true
          description: 可选的子分类名称，必须属于所选一级分类。
          example: 订单管理
        title:
          type: string
          minLength: 1
          maxLength: 160
          description: 建议标题，去除首尾空格后长度不得超过 160。
          example: 增加批量导出发票功能
        description:
          type: string
          minLength: 1
          maxLength: 500
          description: 建议内容，去除首尾空格后长度不得超过 500。
          example: 希望订单页面可以按月份批量导出发票。
        turnstileToken:
          type: string
          minLength: 1
          description: Turnstile Token；备用验证时格式为 fallback:{captchaToken}:{answer}。
          example: 0.AbcDef...
...